seperated components into different files
