+++ apps/excelidraw-frontend/components/Canvas.tsx
@@ -1,5 +1,7 @@
 import { initDraw } from "@/draw";
 import { useEffect, useRef, useState } from "react";
+import { Topbar } from "./Toolbar";
+import { ColorSelector } from "./Sidebar";
 import { IconButton } from "./IconButton";
 import { useRouter } from "next/navigation";
 import {
@@ -23,8 +25,11 @@ import {
     UsersRound,
     ZoomIn,
     ZoomOut,
+    Circle,
+    Pencil,
+    Square,
 } from "lucide-react";
-import { Circle, Pencil, Square } from "lucide-react";
+
 import { Game } from "@/draw/Game";
 import { toast } from "react-toastify";
 
@@ -54,7 +59,7 @@ export type Color =
     | "#a6ffff"
     | "#ffffff";
 
-const colors: Color[] = [
+export const colors: Color[] = [
     "#7a7a7a", // Black
     "#ffa6a6", // Red
     "#a6ffa6", // Green
@@ -308,252 +313,6 @@ export function Canvas({
                     </button>
                 </div>
             </>
-        </>
-    );
-}
-
-function ColorSelector({
-    selectedColor,
-    setSelectedColor,
-    theme,
-    setTheme,
-    title,
-}: {
-    selectedColor: Color;
-    setSelectedColor: (s: Color) => void;
-    theme: Theme;
-    setTheme: (s: Theme) => void;
-    title?: string;
-}) { // Default color is Gray
-    const [showDropdown, setShowDropdown] = useState(false);
-
-    const handleColorSelect = (color: Color) => {
-        setSelectedColor(color);
-        setShowDropdown(false);
-    };
-
-    return (
-        <div className="relative inline-block">
-            <div className="flex gap-2 items-center">
-                <p
-                    className={`${
-                        theme === "rgb(24, 24, 27)"
-                            ? "text-zinc-400"
-                            : "text-zinc-600"
-                    } text-sm`}
-                >
-                    Choose a color:
-                </p>
-                <button
-                    className="p-2 rounded-sm border border-zinc-400"
-                    style={{ backgroundColor: selectedColor }}
-                    onClick={() => setShowDropdown((prev) => !prev)}
-                    title={title}
-                >
-                    {/* The button displays the selected color */}
-                </button>
-            </div>
-            {/* Dropdown to select color */}
-            {showDropdown && (
-                <div
-                    className={`absolute top-full left-0 shadow-md rounded-sm z-10 ${
-                        theme === "rgb(24, 24, 27)" ? "bg-zinc-800" : "bg-zinc-50"
-                    }`}
-                >
-                    <ul className="grid grid-cols-4 gap-2 p-2">
-                        {colors.map((color) => (
-                            <li
-                                key={color}
-                                className="w-4 h-4 rounded-sm cursor-pointer border border-zinc-400"
-                                style={{ backgroundColor: color }}
-                                onClick={() => handleColorSelect(color)} // Handle color selection
-                                title={color.toString()} // Tooltip to show color hex code
-                            >
-                            </li>
-                        ))}
-                    </ul>
-                </div>
-            )}
-        </div>
-    );
-}
-
-function Topbar(
-    {
-        selectedTool,
-        setSelectedTool,
-        selectedColor,
-        setSelectedColor,
-        theme,
-        setTheme,
-    }: {
-        selectedTool: Tool;
-        setSelectedTool: (s: Tool) => void;
-        selectedColor: Color;
-        setSelectedColor: (s: Color) => void;
-        theme: Theme;
-        setTheme: (s: Theme) => void;
-    },
-) {
-    const [isCopied, setIsCopied] = useState(false);
-    const [collaborativeMode, setCollaborativeMode] = useState(false);
-
-    const handleCopy = () => {
-        const currentUrl = window.location.href; // Get the current URL
-        navigator.clipboard.writeText(currentUrl) // Copy URL to clipboard
-            .then(() => {
-                toast.success("URL copied to clipboard!");
-                setIsCopied(true); // Set the copied state to true
-                setCollaborativeMode(true);
-                setTimeout(() => setIsCopied(false), 2000); // Reset after 2 seconds
-            })
-            .catch((err) => console.error("Failed to copy URL:", err));
-    };
-    return (
-        <>
-            {/* Main Toolbar */}
-            <div
-                style={{
-                    position: "fixed",
-                    top: 10,
-                    left: "50%",
-                    transform: "translateX(-50%)",
-                }}
-                className={`${
-                    theme === "rgb(24, 24, 27)" ? "bg-zinc-800" : "bg-white"
-                } flex gap-2 items-center justify-center shadow-lg bg-clip-padding backdrop-filter backdrop-blur-sm bg-opacity-90 rounded-lg px-4 py-1 text-xs font-mono sm:flex-wrap sm:justify-start sm:left-5 sm:top-5`}
-            >
-                {/* Tool Icons */}
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("point");
-                    }}
-                    activated={selectedTool === "point"}
-                    icon={<MousePointer2 size={16} />}
-                    className="hidden sm:inline-block"
-                    title="Pointer"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("select");
-                    }}
-                    activated={selectedTool === "select"}
-                    icon={<SquareDashedMousePointer size={16} />}
-                    className="hidden sm:inline-block"
-                    title="Select"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("hand");
-                    }}
-                    activated={selectedTool === "hand"}
-                    icon={<Hand size={16} />}
-                    title="Grab"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("rect");
-                    }}
-                    activated={selectedTool === "rect"}
-                    icon={<Square size={16} />}
-                    title="Rectangle"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("rhombus");
-                    }}
-                    activated={selectedTool === "rhombus"}
-                    icon={<Diamond size={16} />}
-                    title="Rhombus"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("circle");
-                    }}
-                    activated={selectedTool === "circle"}
-                    icon={<Circle size={16} />}
-                    title="Circle"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("line");
-                    }}
-                    activated={selectedTool === "line"}
-                    icon={<Minus size={16} />}
-                    title="Line"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("arrow");
-                    }}
-                    activated={selectedTool === "arrow"}
-                    icon={<MoveRight size={16} />}
-                    title="Arrow"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("pencil");
-                    }}
-                    activated={selectedTool === "pencil"}
-                    icon={<Pencil size={16} />}
-                    title="Pencil"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("erase");
-                    }}
-                    activated={selectedTool === "erase"}
-                    icon={<Eraser size={16} />}
-                    title="Erase"
-                />
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("text");
-                    }}
-                    activated={selectedTool === "text"}
-                    icon={<TypeOutline size={16} />}
-                    className="hidden sm:inline-block"
-                    title="Text"
-                />
-
-                <span className="opacity-50 text-gray-300">|</span>
-
-                <IconButton
-                    onClick={() => {
-                        setSelectedTool("clear");
-                    }}
-                    activated={selectedTool === "clear"}
-                    icon={<Trash2 size={16} />}
-                    title="Clear canvas"
-                />
-                {/* Collaboration Button */}
-                <button
-                    onClick={handleCopy}
-                    className={`p-2 rounded-md ${
-                        collaborativeMode ? "bg-green-600" : "bg-none"
-                    } transition-colors duration-300`}
-                    title="Collaborative mode"
-                >
-                    <UsersRound
-                        size={16}
-                        className={`${
-                            theme === "rgb(24, 24, 27)"
-                                ? "text-zinc-400"
-                                : "text-zinc-500"
-                        }`}
-                    />
-                </button>
-            </div>
         </>
     );
 }
